Refactor indexer_exceptions -> exceptions
