+++ medusa/indexers/base.py
@@ -11,7 +11,7 @@ import warnings
 from itertools import chain
 from operator import itemgetter
 
-from medusa.indexers.indexer_exceptions import (
+from medusa.indexers.exceptions import (
     IndexerAttributeNotFound,
     IndexerEpisodeNotFound,
     IndexerSeasonNotFound,
