+++ medusa/indexers/indexer_ui.py
@@ -5,7 +5,7 @@
 import logging
 import warnings
 
-from indexer_exceptions import IndexerUserAbort
+from exceptions import IndexerUserAbort
 
 __author__ = 'p0psicles'
 __version__ = '1.0'
