+++ medusa/indexers/tvdbv2/tvdbv2_api.py
@@ -9,9 +9,9 @@ from medusa import app
 from medusa.app import TVDB_API_KEY
 from medusa.helper.metadata import needs_metadata
 from medusa.indexers.base import (Actor, Actors, BaseIndexer)
-from medusa.indexers.indexer_exceptions import (IndexerAuthFailed, IndexerError, IndexerException,
-                                                IndexerShowIncomplete, IndexerShowNotFound,
-                                                IndexerShowNotFoundInLanguage, IndexerUnavailable)
+from medusa.indexers.exceptions import (IndexerAuthFailed, IndexerError, IndexerException,
+                                        IndexerShowIncomplete, IndexerShowNotFound,
+                                        IndexerShowNotFoundInLanguage, IndexerUnavailable)
 from medusa.indexers.indexer_ui import BaseUI, ConsoleUI
 from medusa.indexers.tvdbv2.fallback import PlexFallback
 from medusa.logger.adapters.style import BraceAdapter
