+++ medusa/indexers/tvmaze/tvmaze_api.py
@@ -7,7 +7,7 @@ from collections import OrderedDict
 from time import time
 
 from medusa.indexers.base import (Actor, Actors, BaseIndexer)
-from medusa.indexers.indexer_exceptions import (
+from medusa.indexers.exceptions import (
     IndexerError,
     IndexerException,
     IndexerShowNotFound,
