+++ medusa/metadata/generic.py
@@ -11,8 +11,8 @@ from medusa.helper.exceptions import ex
 from medusa.helper.metadata import get_image
 from medusa.indexers.api import indexerApi
 from medusa.indexers.config import INDEXER_TMDB, INDEXER_TVDBV2, INDEXER_TVMAZE
-from medusa.indexers.indexer_exceptions import (IndexerEpisodeNotFound, IndexerException,
-                                                IndexerSeasonNotFound, IndexerShowNotFound)
+from medusa.indexers.exceptions import (IndexerEpisodeNotFound, IndexerException,
+                                        IndexerSeasonNotFound, IndexerShowNotFound)
 from medusa.logger.adapters.style import BraceAdapter
 
 from requests.exceptions import RequestException
