+++ medusa/server/web/home/handler.py
@@ -60,7 +60,7 @@ from medusa.helper.exceptions import (
 )
 from medusa.indexers.api import indexerApi
 from medusa.indexers.config import INDEXER_TVDBV2
-from medusa.indexers.indexer_exceptions import (
+from medusa.indexers.exceptions import (
     IndexerException,
     IndexerShowNotFoundInLanguage,
 )
