+++ medusa/show_queue.py
@@ -49,7 +49,7 @@ from medusa.helpers import (
 )
 from medusa.helpers.externals import check_existing_shows
 from medusa.indexers.api import indexerApi
-from medusa.indexers.indexer_exceptions import (
+from medusa.indexers.exceptions import (
     IndexerAttributeNotFound,
     IndexerError,
     IndexerException,
