+++ medusa/tv/episode.py
@@ -53,7 +53,7 @@ from medusa.helper.exceptions import (
 from medusa.helper.mappings import NonEmptyDict
 from medusa.indexers.api import indexerApi
 from medusa.indexers.config import indexerConfig
-from medusa.indexers.indexer_exceptions import (
+from medusa.indexers.exceptions import (
     IndexerEpisodeNotFound,
     IndexerError,
     IndexerSeasonNotFound,
